+++ CMakeLists.txt
@@ -3025,7 +3025,6 @@ foreach(_hdr
   include/grpcpp/security/server_credentials.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
-  include/grpcpp/server_builder_impl.h
   include/grpcpp/server_context.h
   include/grpcpp/server_posix.h
   include/grpcpp/support/async_stream.h
@@ -3617,7 +3616,6 @@ foreach(_hdr
   include/grpcpp/security/server_credentials.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
-  include/grpcpp/server_builder_impl.h
   include/grpcpp/server_context.h
   include/grpcpp/server_posix.h
   include/grpcpp/support/async_stream.h
@@ -4581,7 +4579,6 @@ foreach(_hdr
   include/grpcpp/security/server_credentials.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
-  include/grpcpp/server_builder_impl.h
   include/grpcpp/server_context.h
   include/grpcpp/server_posix.h
   include/grpcpp/support/async_stream.h
