+++ Makefile
@@ -5352,7 +5352,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/server_credentials.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
-    include/grpcpp/server_builder_impl.h \
     include/grpcpp/server_context.h \
     include/grpcpp/server_posix.h \
     include/grpcpp/support/async_stream.h \
@@ -5952,7 +5951,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/server_credentials.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
-    include/grpcpp/server_builder_impl.h \
     include/grpcpp/server_context.h \
     include/grpcpp/server_posix.h \
     include/grpcpp/support/async_stream.h \
@@ -6865,7 +6863,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/server_credentials.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
-    include/grpcpp/server_builder_impl.h \
     include/grpcpp/server_context.h \
     include/grpcpp/server_posix.h \
     include/grpcpp/support/async_stream.h \
