+++ gRPC-C++.podspec
@@ -111,7 +111,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/security/server_credentials.h',
                       'include/grpcpp/server.h',
                       'include/grpcpp/server_builder.h',
-                      'include/grpcpp/server_builder_impl.h',
                       'include/grpcpp/server_context.h',
                       'include/grpcpp/server_posix.h',
                       'include/grpcpp/support/async_stream.h',
