+++ include/grpcpp/impl/codegen/completion_queue.h
@@ -41,10 +41,6 @@
 
 struct grpc_completion_queue;
 
-namespace grpc_impl {
-
-class ServerBuilder;
-}
 namespace grpc {
 
 template <class R>
@@ -67,6 +63,7 @@ class ChannelInterface;
 class ClientContext;
 class CompletionQueue;
 class Server;
+class ServerBuilder;
 class ServerContext;
 class ServerInterface;
 
@@ -408,7 +405,7 @@ class ServerCompletionQueue : public CompletionQueue {
         polling_type_(polling_type) {}
 
   grpc_cq_polling_type polling_type_;
-  friend class ::grpc_impl::ServerBuilder;
+  friend class ServerBuilder;
   friend class Server;
 };
