+++ tools/doxygen/Doxyfile.c++
@@ -1001,7 +1001,6 @@ include/grpcpp/security/credentials.h \
 include/grpcpp/security/server_credentials.h \
 include/grpcpp/server.h \
 include/grpcpp/server_builder.h \
-include/grpcpp/server_builder_impl.h \
 include/grpcpp/server_context.h \
 include/grpcpp/server_posix.h \
 include/grpcpp/support/async_stream.h \
